chore: extract duplicate code
